Revert "float64 -> float32"
This reverts commit e5793cacbd3a04e6bf178fbd5c971fb2d650d8c7.

+++ aircon/daikin/daikin04/entry.go
@@ -38,7 +38,7 @@ func (r *daikin04) Generate(e *aircon.Entry) ([]*hex.HexCode, error) {
 	// Temp
 	if e.Mode == "cool" || e.Mode == "heat" {
 		var temp int
-		if t, ok := e.Temp.(float32); ok {
+		if t, ok := e.Temp.(float64); ok {
 			temp = int(t)
 		} else if t, ok := e.Temp.(int); ok {
 			temp = t
